+++ src/sass/layouts/shopping-list/_pagination.scss
@@ -1,0 +1,7 @@
+.js-hidden {
+  visibility: hidden;
+}
+
+.js-display-none {
+  display: none;
+}
